feat(plugins): pull apart multiple plugin files
