--- lua/plugins/enhanced-search.lua
@@ -1,329 +1,0 @@
--- Enhanced project-wide search with live preview
-return {
-  {
-    -- Live grep with preview
-    "nvim-pack/nvim-spectre",
-    config = function()
-      require("spectre").setup({
-        color_devicons = true,
-        open_cmd = "vnew",
-        live_update = false,
-        line_sep_start = "┌-----------------------------------------",
-        result_padding = "¦  ",
-        line_sep = "└-----------------------------------------",
-        highlight = {
-          ui = "String",
-          search = "DiffChange",
-          replace = "DiffDelete",
-        },
-        mapping = {
-          ["toggle_line"] = {
-            map = "dd",
-            cmd = "<cmd>lua require('spectre').toggle_line()<CR>",
-            desc = "toggle current item",
-          },
-          ["enter_file"] = {
-            map = "<cr>",
-            cmd = "<cmd>lua require('spectre.actions').select_entry()<CR>",
-            desc = "goto current file",
-          },
-          ["send_to_qf"] = {
-            map = "<leader>q",
-            cmd = "<cmd>lua require('spectre.actions').send_to_qf()<CR>",
-            desc = "send all item to quickfix",
-          },
-          ["replace_cmd"] = {
-            map = "<leader>c",
-            cmd = "<cmd>lua require('spectre.actions').replace_cmd()<CR>",
-            desc = "input replace vim command",
-          },
-          ["show_option_menu"] = {
-            map = "<leader>o",
-            cmd = "<cmd>lua require('spectre').show_options()<CR>",
-            desc = "show option",
-          },
-          ["run_current_replace"] = {
-            map = "<leader>rc",
-            cmd = "<cmd>lua require('spectre.actions').run_current_replace()<CR>",
-            desc = "replace current line",
-          },
-          ["run_replace"] = {
-            map = "<leader>R",
-            cmd = "<cmd>lua require('spectre.actions').run_replace()<CR>",
-            desc = "replace all",
-          },
-          ["change_view_mode"] = {
-            map = "<leader>v",
-            cmd = "<cmd>lua require('spectre').change_view()<CR>",
-            desc = "change result view mode",
-          },
-          ["change_replace_sed"] = {
-            map = "trs",
-            cmd = "<cmd>lua require('spectre').change_engine_replace('sed')<CR>",
-            desc = "use sed to replace",
-          },
-          ["change_replace_oxi"] = {
-            map = "tro",
-            cmd = "<cmd>lua require('spectre').change_engine_replace('oxi')<CR>",
-            desc = "use oxi to replace",
-          },
-          ["toggle_live_update"] = {
-            map = "tu",
-            cmd = "<cmd>lua require('spectre').toggle_live_update()<CR>",
-            desc = "update change when vim write file.",
-          },
-          ["toggle_ignore_case"] = {
-            map = "ti",
-            cmd = "<cmd>lua require('spectre').change_options('ignore-case')<CR>",
-            desc = "toggle ignore case",
-          },
-          ["toggle_ignore_hidden"] = {
-            map = "th",
-            cmd = "<cmd>lua require('spectre').change_options('hidden')<CR>",
-            desc = "toggle search hidden",
-          },
-          ["resume_last_search"] = {
-            map = "<leader>l",
-            cmd = "<cmd>lua require('spectre').resume_last_search()<CR>",
-            desc = "resume last search before close",
-          },
-        },
-        find_engine = {
-          ["rg"] = {
-            cmd = "rg",
-            args = {
-              "--color=never",
-              "--no-heading",
-              "--with-filename",
-              "--line-number",
-              "--column",
-            },
-            options = {
-              ["ignore-case"] = {
-                value = "--ignore-case",
-                icon = "[I]",
-                desc = "ignore case",
-              },
-              ["hidden"] = {
-                value = "--hidden",
-                desc = "hidden file",
-                icon = "[H]",
-              },
-            },
-          },
-        },
-        replace_engine = {
-          ["sed"] = {
-            cmd = "sed",
-            args = nil,
-            options = {
-              ["ignore-case"] = {
-                value = "--ignore-case",
-                icon = "[I]",
-                desc = "ignore case",
-              },
-            },
-          },
-        },
-        default = {
-          find = {
-            cmd = "rg",
-            options = { "ignore-case" },
-          },
-          replace = {
-            cmd = "sed",
-          },
-        },
-        replace_vim_cmd = "cdo",
-        is_open = false,
-        is_insert_mode = false,
-      })
-
-      -- Keymaps
-      vim.keymap.set("n", "<leader>S", '<cmd>lua require("spectre").toggle()<CR>', { desc = "Toggle Spectre" })
-      vim.keymap.set(
-        "n",
-        "<leader>sw",
-        '<cmd>lua require("spectre").open_visual({select_word=true})<CR>',
-        { desc = "Search current word" }
-      )
-      vim.keymap.set(
-        "v",
-        "<leader>sw",
-        '<esc><cmd>lua require("spectre").open_visual()<CR>',
-        { desc = "Search current word" }
-      )
-      vim.keymap.set(
-        "n",
-        "<leader>sp",
-        '<cmd>lua require("spectre").open_file_search({select_word=true})<CR>',
-        { desc = "Search on current file" }
-      )
-    end,
-  },
-  {
-    -- Enhanced telescope with more features
-    "nvim-telescope/telescope.nvim",
-    dependencies = {
-      "nvim-lua/plenary.nvim",
-      "nvim-telescope/telescope-fzf-native.nvim",
-
-      "nvim-telescope/telescope-file-browser.nvim",
-    },
-    config = function()
-      local telescope = require("telescope")
-      local actions = require("telescope.actions")
-
-      telescope.setup({
-        defaults = {
-          prompt_prefix = "🔍 ",
-          selection_caret = "❯ ",
-          path_display = { "truncate" },
-          file_ignore_patterns = {
-            "^.git/",
-            "^.svn/",
-            "^.hg/",
-            "^CVS/",
-            "^.DS_Store",
-            "node_modules",
-            ".pytest_cache",
-            "__pycache__",
-            ".terraform",
-            "*.tfstate*",
-            ".elixir_ls",
-            "_build",
-            "deps",
-          },
-          layout_config = {
-            horizontal = {
-              prompt_position = "top",
-              preview_width = 0.55,
-              results_width = 0.8,
-            },
-            vertical = {
-              mirror = false,
-            },
-            width = 0.87,
-            height = 0.80,
-            preview_cutoff = 120,
-          },
-          mappings = {
-            i = {
-              ["<C-n>"] = actions.cycle_history_next,
-              ["<C-p>"] = actions.cycle_history_prev,
-              ["<C-j>"] = actions.move_selection_next,
-              ["<C-k>"] = actions.move_selection_previous,
-              ["<CR>"] = actions.select_default,
-              ["<C-x>"] = actions.select_horizontal,
-              ["<C-v>"] = actions.select_vertical,
-              ["<C-t>"] = actions.select_tab,
-              ["<C-u>"] = actions.preview_scrolling_up,
-              ["<C-d>"] = actions.preview_scrolling_down,
-              ["<PageUp>"] = actions.results_scrolling_up,
-              ["<PageDown>"] = actions.results_scrolling_down,
-              ["<Tab>"] = actions.toggle_selection + actions.move_selection_worse,
-              ["<S-Tab>"] = actions.toggle_selection + actions.move_selection_better,
-              ["<C-q>"] = actions.send_to_qflist + actions.open_qflist,
-              ["<M-q>"] = actions.send_selected_to_qflist + actions.open_qflist,
-              ["<C-l>"] = actions.complete_tag,
-              ["<C-/>"] = actions.which_key,
-            },
-            n = {
-              ["<esc>"] = actions.close,
-              ["<CR>"] = actions.select_default,
-              ["<C-x>"] = actions.select_horizontal,
-              ["<C-v>"] = actions.select_vertical,
-              ["<C-t>"] = actions.select_tab,
-              ["<Tab>"] = actions.toggle_selection + actions.move_selection_worse,
-              ["<S-Tab>"] = actions.toggle_selection + actions.move_selection_better,
-              ["<C-q>"] = actions.send_to_qflist + actions.open_qflist,
-              ["<M-q>"] = actions.send_selected_to_qflist + actions.open_qflist,
-              ["j"] = actions.move_selection_next,
-              ["k"] = actions.move_selection_previous,
-              ["H"] = actions.move_to_top,
-              ["M"] = actions.move_to_middle,
-              ["L"] = actions.move_to_bottom,
-              ["<Down>"] = actions.move_selection_next,
-              ["<Up>"] = actions.move_selection_previous,
-              ["gg"] = actions.move_to_top,
-              ["G"] = actions.move_to_bottom,
-              ["<C-u>"] = actions.preview_scrolling_up,
-              ["<C-d>"] = actions.preview_scrolling_down,
-              ["<PageUp>"] = actions.results_scrolling_up,
-              ["<PageDown>"] = actions.results_scrolling_down,
-              ["?"] = actions.which_key,
-            },
-          },
-        },
-        pickers = {
-          find_files = {
-            theme = "dropdown",
-            hidden = true,
-            find_command = { "rg", "--files", "--hidden", "--glob", "!.git/*" },
-          },
-          live_grep = {
-            additional_args = function(opts)
-              return { "--hidden", "--glob", "!.git/*" }
-            end,
-          },
-          grep_string = {
-            additional_args = function(opts)
-              return { "--hidden", "--glob", "!.git/*" }
-            end,
-          },
-          buffers = {
-            theme = "dropdown",
-            previewer = false,
-            initial_mode = "normal",
-          },
-          colorscheme = {
-            enable_preview = true,
-          },
-        },
-        extensions = {
-          fzf = {
-            fuzzy = true,
-            override_generic_sorter = true,
-            override_file_sorter = true,
-            case_mode = "smart_case",
-          },
-
-          file_browser = {
-            theme = "ivy",
-            hijack_netrw = true,
-            mappings = {
-              ["i"] = {
-                -- your custom insert mode mappings
-              },
-              ["n"] = {
-                -- your custom normal mode mappings
-              },
-            },
-          },
-        },
-      })
-
-      -- Load extensions
-      telescope.load_extension("fzf")
-      telescope.load_extension("file_browser")
-
-      -- Enhanced keymaps
-      vim.keymap.set("n", "<leader>ff", "<cmd>Telescope find_files<cr>", { desc = "Find files" })
-      vim.keymap.set("n", "<leader>fg", "<cmd>Telescope live_grep<cr>", { desc = "Live grep" })
-
-      vim.keymap.set("n", "<leader>fb", "<cmd>Telescope buffers<cr>", { desc = "Buffers" })
-      vim.keymap.set("n", "<leader>fh", "<cmd>Telescope help_tags<cr>", { desc = "Help tags" })
-      vim.keymap.set("n", "<leader>fo", "<cmd>Telescope oldfiles<cr>", { desc = "Recent files" })
-      vim.keymap.set("n", "<leader>fc", "<cmd>Telescope colorscheme<cr>", { desc = "Colorschemes" })
-      vim.keymap.set("n", "<leader>fr", "<cmd>Telescope resume<cr>", { desc = "Resume last search" })
-      vim.keymap.set("n", "<leader>fk", "<cmd>Telescope keymaps<cr>", { desc = "Keymaps" })
-      vim.keymap.set("n", "<leader>fd", ":Telescope file_browser<CR>", { desc = "File browser" })
-      vim.keymap.set(
-        "n",
-        "<leader>fD",
-        ":Telescope file_browser path=%:p:h select_buffer=true<CR>",
-        { desc = "File browser (current dir)" }
-      )
-    end,
-  },
-}
