+++ lua/plugins/markdown-preview.lua
@@ -1,0 +1,30 @@
+-- Markdown preview
+return {
+  "iamcco/markdown-preview.nvim",
+  cmd = { "MarkdownPreviewToggle", "MarkdownPreview", "MarkdownPreviewStop" },
+  ft = { "markdown" },
+  build = function()
+    vim.fn["mkdp#util#install"]()
+  end,
+  config = function()
+    vim.g.mkdp_auto_start = 0
+    vim.g.mkdp_auto_close = 1
+    vim.g.mkdp_refresh_slow = 0
+    vim.g.mkdp_command_for_global = 0
+    vim.g.mkdp_open_to_the_world = 0
+    vim.g.mkdp_open_ip = ""
+    vim.g.mkdp_browser = ""
+    vim.g.mkdp_echo_preview_url = 0
+    vim.g.mkdp_browserfunc = ""
+    vim.g.mkdp_page_title = "「${name}」"
+    vim.g.mkdp_filetypes = { "markdown" }
+    vim.g.mkdp_theme = "dark"
+    vim.g.mkdp_combine_preview = 0
+    vim.g.mkdp_combine_preview_auto_refresh = 1
+
+    -- Keymaps
+    vim.keymap.set("n", "<leader>mp", "<cmd>MarkdownPreview<cr>", { desc = "Markdown preview" })
+    vim.keymap.set("n", "<leader>mP", "<cmd>MarkdownPreviewStop<cr>", { desc = "Stop markdown preview" })
+    vim.keymap.set("n", "<leader>mt", "<cmd>MarkdownPreviewToggle<cr>", { desc = "Toggle markdown preview" })
+  end,
+}
