+++ lua/plugins/nvim-ufo.lua
@@ -1,0 +1,48 @@
+-- Advanced folding with treesitter
+return {
+  "kevinhwang91/nvim-ufo",
+  dependencies = "kevinhwang91/promise-async",
+  config = function()
+    vim.o.foldcolumn = "1"
+    vim.o.foldlevel = 99
+    vim.o.foldlevelstart = 99
+    vim.o.foldenable = true
+
+    require("ufo").setup({
+      provider_selector = function(bufnr, filetype, buftype)
+        return { "treesitter", "indent" }
+      end,
+      open_fold_hl_timeout = 150,
+      close_fold_kinds_for_ft = {
+        default = { "imports", "comment" },
+        json = { "array" },
+        c = { "comment", "region" },
+      },
+      preview = {
+        win_config = {
+          border = { "", "─", "", "", "", "─", "", "" },
+          winhighlight = "Normal:Folded",
+          winblend = 0,
+        },
+        mappings = {
+          scrollU = "<C-u>",
+          scrollD = "<C-d>",
+          jumpTop = "[",
+          jumpBot = "]",
+        },
+      },
+    })
+
+    -- Keymaps
+    vim.keymap.set("n", "zR", require("ufo").openAllFolds, { desc = "Open all folds" })
+    vim.keymap.set("n", "zM", require("ufo").closeAllFolds, { desc = "Close all folds" })
+    vim.keymap.set("n", "zr", require("ufo").openFoldsExceptKinds, { desc = "Open folds except kinds" })
+    vim.keymap.set("n", "zm", require("ufo").closeFoldsWith, { desc = "Close folds with" })
+    vim.keymap.set("n", "K", function()
+      local winid = require("ufo").peekFoldedLinesUnderCursor()
+      if not winid then
+        vim.lsp.buf.hover()
+      end
+    end, { desc = "Peek fold or hover" })
+  end,
+}
