+++ lua/plugins/targets.lua
@@ -1,0 +1,8 @@
+-- Enhanced text objects
+return {
+  "wellle/targets.vim",
+  -- Provides enhanced text objects like:
+  -- cin) - change inside next parentheses
+  -- da, - delete around comma
+  -- ci2) - change inside 2nd parentheses
+}
