+++ lua/plugins/virt-column.lua
@@ -1,0 +1,11 @@
+-- Color columns and line length guides
+return {
+  "lukas-reineke/virt-column.nvim",
+  config = function()
+    require("virt-column").setup({
+      virtcolumn = "80,120",
+      char = "┃",
+      highlight = { "NonText" },
+    })
+  end,
+}
